+++ models/orange_pasta/configs/config_common.py
@@ -6,10 +6,6 @@ def get_common_config():
         "queryset": "fatalities003_pgm_baseline",
         "data_train": "baseline",
         "level": "pgm",
-        
         'steps': [*range(1, 36 + 1, 1)],
-        'calib_partitioner_dict': {"train": (121, 396), "predict": (397, 444)},
-        'test_partitioner_dict': {"train": (121, 444), "predict": (445, 492)},
-        'future_partitioner_dict': {"train": (121, 492), "predict": (493, 504)},
     }
     return common_config
