+++ models/orange_pasta/src/forecasting/generate_forecast.py
@@ -2,23 +2,22 @@ import pandas as pd
 from pathlib import Path
 
 import sys
-src_path = f"{Path(__file__).parent.parent}"
-sys.path.append(str(src_path)+"/utils")
+pipeline_path = f"{Path(__file__).parent.parent.parent.parent.parent}"
+sys.path.append(str(pipeline_path)+"/common_utils")
+from common_utils.set_path import get_artifacts_path, get_data_path
 
-# Remove this part after packaging views_stepshift
-current_file_path = Path(__file__).resolve()
-root_path = current_file_path.parent.parent.parent.parent.parent
-sys.path.append(str(root_path))
 
-from utils import get_artifacts_path, get_data_path
-
-def forecast():
+def forecast(common_config):
     print('Predicting...')
 
-    stepshifter_model = pd.read_pickle(get_artifacts_path("future"))
-    dataset = pd.read_parquet(get_data_path("raw"))
+    stepshifter_model = pd.read_pickle(get_artifacts_path(common_config["name"], "forecast"))
+    dataset = pd.read_parquet(get_data_path(common_config["name"], "raw"))
 
-    # Predictions for the future (NaN values exist as it uses data from the last point in time in the selected partition):
-    predictions = stepshifter_model.future_predict("future", "predict", dataset, keep_specific=True)
+    predictions = stepshifter_model.predict("forecast", "predict", dataset)
+
+    pred_cols = [f"step_pred_{str(i)}" for i in common_config["steps"]]
+    predictions = predictions[pred_cols]
+
+    predictions.to_parquet(get_data_path(common_config["name"], "generated"))
 
     return predictions
