+++ models/yellow_pikachu/configs/config_common.py
@@ -6,12 +6,6 @@ def get_common_config():
         "queryset": "fatalities003_pgm_conflict_treelag",
         "data_train": "conflict_treelag",
         "level": "pgm",
-        
         'steps': [*range(1, 36 + 1, 1)],
-        'calib_partitioner_dict': {"train": (121, 396), "predict": (397, 444)},
-        'test_partitioner_dict': {"train": (121, 444), "predict": (445, 492)},
-        'future_partitioner_dict': {"train": (121, 492), "predict": (493, 504)},
-        'FutureStart': 508,
-        'force_retrain': False
     }
     return common_config
