+++ models/yellow_pikachu/src/training/train_model.py
@@ -2,48 +2,43 @@ from pathlib import Path
 import pandas as pd
 
 import sys
-src_path = f"{Path(__file__).parent.parent}"
-sys.path.append(str(src_path)+"/utils")
-
-# Remove this part after packaging views_stepshift
-current_file_path = Path(__file__).resolve()
-root_path = current_file_path.parent.parent.parent.parent.parent
-sys.path.append(str(root_path))
+pipeline_path = f"{Path(__file__).parent.parent.parent.parent.parent}"
+sys.path.append(str(pipeline_path)+"/common_utils")
 
 from xgboost import XGBRegressor
 
-from utils import get_artifacts_path, get_data_path
+from common_utils.set_path import get_artifacts_path, get_data_path
 from views_partitioning.data_partitioner import DataPartitioner
 from views_forecasts.extensions import *
 from stepshift.views import StepshiftedModels
-from views_stepshift.run import ViewsRun
+from common_utils.views_stepshift.run import ViewsRun
 
 
 def train(common_config, para_config):
     print("Training...")
     if not common_config["sweep"]:
         model = globals()[common_config["algorithm"]](**para_config)
-        dataset = pd.read_parquet(get_data_path("raw"))
+        dataset = pd.read_parquet(get_data_path(common_config["name"], "raw"))
         # Train partition
         try:
-            stepshifter_model_calib = pd.read_pickle(get_artifacts_path("calib"))
+            stepshifter_model_calib = pd.read_pickle(get_artifacts_path(common_config["name"], "calib"))
         except:
             stepshifter_model_calib = stepshift_training(common_config, "calib", model, dataset)
-            stepshifter_model_calib.save(get_artifacts_path("calib"))
+            stepshifter_model_calib.save(get_artifacts_path(common_config["name"], "calib"))
 
         # Test partition
         try:
-            stepshifter_model_test = pd.read_pickle(get_artifacts_path("test"))
+            stepshifter_model_test = pd.read_pickle(get_artifacts_path(common_config["name"], "test"))
         except:
             stepshifter_model_test = stepshift_training(common_config, "test", model, dataset)
-            stepshifter_model_test.save(get_artifacts_path("test"))
+            stepshifter_model_test.save(get_artifacts_path(common_config["name"], "test"))
 
         # Future partition
         try:
-            stepshifter_model_future = pd.read_pickle(get_artifacts_path("future"))
+            stepshifter_model_future = pd.read_pickle(get_artifacts_path(common_config["name"], "forecast"))
         except:
-            stepshifter_model_future = stepshift_training(common_config, "future", model, dataset)
-            stepshifter_model_future.save(get_artifacts_path("future"))  
+            stepshifter_model_future = stepshift_training(common_config, "forecast", model, dataset)
+            stepshifter_model_future.save(get_artifacts_path(common_config["name"], "forecast")) 
 
 
 def stepshift_training(common_config, partition_name, model, dataset):
